refactor: separate non-native field from api wrapper
