+++ internal/stats/snippet.go
@@ -79,8 +79,8 @@ func initSnippets() {
 		mimc.Write(newVariable())
 		_ = mimc.Sum()
 	})
-	registerSnippet("math/emulated/secp256k1_32", func(api frontend.API, newVariable func() frontend.Variable) {
-		secp256k1, _ := emulated.NewField[emulated.Secp256k1](api)
+	registerSnippet("math/emulated/secp256k1_64", func(api frontend.API, newVariable func() frontend.Variable) {
+		secp256k1, _ := emulated.NewAPI[emulated.Secp256k1](api)
 
 		newElement := func() emulated.Element[emulated.Secp256k1] {
 			r := emulated.NewElement[emulated.Secp256k1](nil)
